Add documentation

+++ SegaAMFileLib/CryptHash/SegaCRC32.cs
@@ -3,13 +3,27 @@ using Haruka.Arcade.SegaAMFileLib.Debugging;
 using Microsoft.Extensions.Logging;
 
 namespace Haruka.Arcade.SegaAMFileLib.CryptHash {
+    
+    /// <summary>
+    /// Implementation of the CRC32 algorithm that SEGA uses.
+    /// </summary>
     public static class SegaCrc32 {
         private static readonly ILogger LOG = Logging.Factory.CreateLogger(nameof(SegaCrc32));
 
+        /// <summary>
+        /// Calculates the CRC32 for the given byte array. Remember that if the CRC field is part of this array, it should be zeroed out.
+        /// </summary>
+        /// <param name="data">The byte array to use.</param>
+        /// <returns>The computed CRC32 checksum.</returns>
         public static uint CalcCrc32(byte[] data) {
             return new Crc32Managed().GetCrc32(data);
         }
 
+        /// <summary>
+        /// Calculates the CRC32 for the given byte array, and writes it into the first 4 bytes of the given byte array.
+        /// </summary>
+        /// <param name="struc"></param>
+        /// <returns></returns>
         public static byte[] WriteCrcIntoFirst4Bytes(byte[] struc) {
             byte[] crcableBytes = new byte[struc.Length - 4];
             Array.Copy(struc, 0 + sizeof(uint), crcableBytes, 0, struc.Length - sizeof(uint));
@@ -32,7 +46,7 @@ namespace Haruka.Arcade.SegaAMFileLib.CryptHash {
             uint crc = 0xFFFFFFFF;
             for (int i = 0; i < data.Length; i++) {
                 byte index = (byte)(((crc) & 0xFF) ^ data[i]);
-                crc = (uint)((crc >> 8) ^ table[index]);
+                crc = (crc >> 8) ^ table[index];
             }
 
             return ~crc;
@@ -45,7 +59,7 @@ namespace Haruka.Arcade.SegaAMFileLib.CryptHash {
                 uint temp = i;
                 for (int j = 8; j > 0; j--) {
                     if ((temp & 1) == 1) {
-                        temp = (uint)((temp >> 1) ^ poly);
+                        temp = (temp >> 1) ^ poly;
                     } else {
                         temp >>= 1;
                     }
